+++ xcell/mappers/mapper_eBOSSQSO.py
@@ -1,7 +1,8 @@
 from .mapper_base import MapperBase
-from .utils import get_map_from_points
+from .mapper_base import MapperSDSS
 from astropy.io import fits
 from astropy.table import Table, vstack
+from .utils import get_map_from_points
 import numpy as np
 import healpy as hp
 import pymaster as nmt
@@ -20,7 +21,6 @@ class MappereBOSSQSO(MapperSDSS):
            'mask_name': 'mask_QSO_NGC_1'}
         """
         self._get_defaults(config)
-
         self.cats = {'data': None, 'random': None}
 
         self.z_arr_dim = config.get('z_arr_dim', 50)
@@ -35,22 +35,6 @@ class MappereBOSSQSO(MapperSDSS):
         self.delta_map = None
         self.nl_coupled = None
         self.mask = None
-
-    def get_catalog(self, mod='data'):
-        if mod == 'data':
-            data_file = self.config['data_catalogs']
-        else:
-            data_file = self.config['random_catalogs']
-
-        if self.cats[mod] is None:
-            cats = []
-            for file in data_file:
-                if not os.path.isfile(file):
-                    raise ValueError(f"File {file} not found")
-                with fits.open(file) as f:
-                    cats.append(self._bin_z(Table.read(f)))
-            self.cats[mod] = vstack(cats)
-        return self.cats[mod]
 
     def _bin_z(self, cat):
         return cat[(cat['Z'] >= self.z_edges[0]) &
@@ -68,6 +52,30 @@ class MappereBOSSQSO(MapperSDSS):
         z_dz = z + dz
         sel = z_dz >= 0
         return np.array([z_dz[sel], nz[sel]])
+    
+    def _get_w(self, mod='data'):
+        if self.ws[mod] is None:
+            cat = self.get_catalog(mod=mod)
+            cat_SYSTOT = np.array(cat['WEIGHT_SYSTOT'])
+            cat_CP = np.array(cat['WEIGHT_CP'])
+            cat_NOZ = np.array(cat['WEIGHT_NOZ'])
+            self.ws[mod] = cat_SYSTOT*cat_CP*cat_NOZ  # FKP left out
+        return self.ws[mod]
+    
+    def get_mask(self):
+        if self.mask is None:
+            cat_random = self.get_catalog(mod='random')
+            w_random = self._get_w(mod='random')
+            alpha = self._get_alpha()
+            self.mask = get_map_from_points(cat_random,
+                                            self.nside_mask,
+                                            w=w_random)
+            self.mask *= alpha
+            # Account for different pixel areas
+            area_ratio = (self.nside_mask/self.nside)**2
+            self.mask = area_ratio * hp.ud_grade(self.mask,
+                                                 nside_out=self.nside)
+        return self.mask
 
 
     def get_dtype(self):
